refactor: KEEP-1351 extract duplicate getAdminFetchHeaders to shared module
The same function was duplicated in auth.ts and invitations.ts.
Extracted to admin-fetch.ts and imported from both.
