+++ tests/e2e/playwright/utils/admin-fetch.ts
@@ -1,0 +1,10 @@
+export function getAdminFetchHeaders(): Record<string, string> {
+  const headers: Record<string, string> = {
+    Authorization: `Bearer ${process.env.TEST_API_KEY}`,
+  };
+  if (process.env.CF_ACCESS_CLIENT_ID && process.env.CF_ACCESS_CLIENT_SECRET) {
+    headers["CF-Access-Client-Id"] = process.env.CF_ACCESS_CLIENT_ID;
+    headers["CF-Access-Client-Secret"] = process.env.CF_ACCESS_CLIENT_SECRET;
+  }
+  return headers;
+}
