+++ tests/e2e/playwright/utils/auth.ts
@@ -1,17 +1,7 @@
 import type { Page } from "@playwright/test";
 import { expect } from "@playwright/test";
 import postgres from "postgres";
-
-function getAdminFetchHeaders(): Record<string, string> {
-  const headers: Record<string, string> = {
-    Authorization: `Bearer ${process.env.TEST_API_KEY}`,
-  };
-  if (process.env.CF_ACCESS_CLIENT_ID && process.env.CF_ACCESS_CLIENT_SECRET) {
-    headers["CF-Access-Client-Id"] = process.env.CF_ACCESS_CLIENT_ID;
-    headers["CF-Access-Client-Secret"] = process.env.CF_ACCESS_CLIENT_SECRET;
-  }
-  return headers;
-}
+import { getAdminFetchHeaders } from "./admin-fetch";
 
 /**
  * Sign up a new user and navigate to verification view.
